Added basic interactivity and audio visualization
I have some rudimentary code in there that allows me to play pause the music with the DOM.

I also used ThreeJS's audio playback and analysis tools to drive the scale of the red cube. This is not going to be the final animation, but I have the framework and data to drive all sorts of 3D objects with the audio now.